feat: Add tests, coverage, and improve project structure
This commit introduces several improvements to the `dbtool` project:

- **Project Refactoring:** The codebase has been refactored into a library crate (`src/lib.rs`) and a binary crate (`src/main.rs`). This improves modularity, separates concerns, and makes the code more testable.

- **Added Integration Tests:** A new integration test suite has been added in the `tests/` directory. These tests cover the main functionality of the CLI, including the `init`, `validate`, `plan`, `apply`, and `destroy` commands.
  - **Note:** The tests that interact with the database require a running PostgreSQL instance. They are configured to connect to `postgres://postgres:postgres@127.0.0.1:5432/postgres`.

- **Test Coverage Analysis:** The project is now set up to use `cargo-tarpaulin` for code coverage analysis. A baseline coverage report can be generated, although it will be low until the database-dependent tests are run in a suitable environment.

- **`Cargo.toml` Correction:** The Rust edition in `Cargo.toml` has been corrected from `2024` to `2021`.

- **README Update (Attempted):** An attempt was made to update the `README.md` to be bilingual (English/Portuguese), add a section on testing, and include a coverage badge. However, writing the new `README.md` file failed due to a tool error. The new `README.md` content is ready and can be added manually.
