+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "dbtool"
 version = "0.1.0"
-edition = "2024"
+edition = "2021"
 
 [dependencies]
 clap = { version = "4.0", features = ["derive"] }
@@ -12,3 +12,9 @@ tokio = { version = "1.0", features = ["full"] }
 tracing = "0.1"
 tracing-subscriber = "0.3"
 anyhow = "1.0"
+
+[dev-dependencies]
+assert_cmd = "2.0"
+sqlx = { version = "0.7", features = ["runtime-tokio", "postgres"] }
+tokio = { version = "1.0", features = ["full"] }
+predicates = "2.1"
