+++ create_test_table_apply_destroy.sql
@@ -1,0 +1,1 @@
+CREATE TABLE test_table_apply_destroy (id SERIAL PRIMARY KEY);
