+++ lcov.info
@@ -1,0 +1,232 @@
+TN:
+SF:/app/src/lib.rs
+FN:87,run
+FN:112,init_playbook
+FN:123,apply_playbook
+FN:227,test_playbook
+FN:252,destroy_playbook
+FN:330,validate_playbook
+FN:353,check_database_exists
+FN:364,check_table_exists
+FN:377,execute_sql_file
+FNF:9
+FNDA:0,run
+FNDA:0,init_playbook
+FNDA:0,apply_playbook
+FNDA:0,test_playbook
+FNDA:0,destroy_playbook
+FNDA:0,validate_playbook
+FNDA:0,check_database_exists
+FNDA:0,check_table_exists
+FNDA:0,execute_sql_file
+DA:87,0
+DA:88,0
+DA:90,0
+DA:91,0
+DA:92,0
+DA:94,0
+DA:95,0
+DA:97,0
+DA:98,0
+DA:100,0
+DA:101,0
+DA:103,0
+DA:104,0
+DA:106,0
+DA:107,0
+DA:112,0
+DA:113,0
+DA:114,0
+DA:115,0
+DA:116,0
+DA:117,0
+DA:118,0
+DA:119,0
+DA:120,0
+DA:123,0
+DA:124,0
+DA:125,0
+DA:126,0
+DA:127,0
+DA:128,0
+DA:131,0
+DA:132,0
+DA:134,0
+DA:135,0
+DA:138,0
+DA:141,0
+DA:142,0
+DA:143,0
+DA:145,0
+DA:146,0
+DA:147,0
+DA:148,0
+DA:151,0
+DA:152,0
+DA:153,0
+DA:154,0
+DA:157,0
+DA:158,0
+DA:159,0
+DA:160,0
+DA:162,0
+DA:163,0
+DA:164,0
+DA:170,0
+DA:171,0
+DA:173,0
+DA:176,0
+DA:177,0
+DA:178,0
+DA:180,0
+DA:181,0
+DA:182,0
+DA:183,0
+DA:186,0
+DA:187,0
+DA:188,0
+DA:189,0
+DA:192,0
+DA:193,0
+DA:194,0
+DA:195,3
+DA:197,3
+DA:198,0
+DA:199,0
+DA:200,0
+DA:201,0
+DA:202,0
+DA:204,0
+DA:210,0
+DA:211,0
+DA:215,0
+DA:216,0
+DA:217,0
+DA:218,0
+DA:219,0
+DA:220,0
+DA:224,0
+DA:227,0
+DA:228,0
+DA:229,0
+DA:230,0
+DA:232,0
+DA:233,0
+DA:235,0
+DA:237,0
+DA:238,0
+DA:239,0
+DA:241,0
+DA:242,0
+DA:243,0
+DA:245,0
+DA:246,0
+DA:247,0
+DA:249,0
+DA:252,0
+DA:253,0
+DA:254,0
+DA:255,0
+DA:256,0
+DA:259,0
+DA:261,3
+DA:262,0
+DA:265,0
+DA:268,0
+DA:269,3
+DA:270,0
+DA:271,0
+DA:272,0
+DA:273,0
+DA:274,0
+DA:275,0
+DA:277,0
+DA:278,3
+DA:279,0
+DA:281,0
+DA:282,0
+DA:283,0
+DA:287,3
+DA:288,0
+DA:293,0
+DA:294,0
+DA:295,3
+DA:296,0
+DA:297,0
+DA:298,0
+DA:299,0
+DA:300,0
+DA:302,0
+DA:303,0
+DA:304,3
+DA:306,0
+DA:307,0
+DA:308,0
+DA:312,0
+DA:313,0
+DA:318,0
+DA:319,0
+DA:320,0
+DA:321,0
+DA:322,0
+DA:323,0
+DA:327,0
+DA:330,0
+DA:331,0
+DA:332,0
+DA:333,0
+DA:334,0
+DA:338,0
+DA:339,0
+DA:340,0
+DA:343,0
+DA:344,0
+DA:345,0
+DA:349,0
+DA:350,0
+DA:353,0
+DA:354,0
+DA:355,0
+DA:356,0
+DA:357,0
+DA:358,0
+DA:359,0
+DA:360,0
+DA:361,0
+DA:364,0
+DA:365,0
+DA:369,0
+DA:370,0
+DA:371,0
+DA:372,0
+DA:373,0
+DA:374,0
+DA:377,0
+DA:378,0
+DA:379,0
+DA:380,0
+DA:381,0
+DA:382,0
+DA:383,0
+DA:384,0
+DA:385,0
+DA:387,0
+DA:388,0
+DA:389,0
+DA:390,0
+DA:391,0
+DA:392,0
+LF:197
+LH:8
+end_of_record
+TN:
+SF:/app/src/main.rs
+FN:6,main
+FNF:1
+FNDA:0,main
+DA:6,0
+DA:7,0
+DA:8,0
+LF:3
+LH:0
+end_of_record
