+++ test_db_apply_destroy.yml
@@ -1,0 +1,5 @@
+
+databases:
+  - name: test_db_apply_destroy
+    if_not_exists: create_test_db_apply_destroy.sql
+tables: []
