+++ test_table_apply_destroy.yml
@@ -1,0 +1,6 @@
+
+databases: []
+tables:
+  - database: postgres
+    name: test_table_apply_destroy
+    if_not_exists: create_test_table_apply_destroy.sql
